+++ apps/aides/admin/__init__.py
@@ -1,0 +1,2 @@
+from .aide import *  # noqa: F403
+from .other_models import *  # noqa: F403
